+++ src/Extension/Mention/MentionParser.php
@@ -27,7 +27,7 @@ final class MentionParser implements InlineParserInterface
      *
      * @psalm-readonly
      */
-    private $symbol;
+    private $prefix;
 
     /**
      * @var string
@@ -43,23 +43,23 @@ final class MentionParser implements InlineParserInterface
      */
     private $mentionGenerator;
 
-    public function __construct(string $symbol, string $mentionRegex, MentionGeneratorInterface $mentionGenerator)
+    public function __construct(string $prefix, string $mentionRegex, MentionGeneratorInterface $mentionGenerator)
     {
-        $this->symbol           = $symbol;
+        $this->prefix           = $prefix;
         $this->mentionRegex     = $mentionRegex;
         $this->mentionGenerator = $mentionGenerator;
     }
 
     public function getMatchDefinition(): InlineParserMatch
     {
-        return InlineParserMatch::string($this->symbol);
+        return InlineParserMatch::string($this->prefix);
     }
 
     public function parse(string $match, InlineParserContext $inlineContext): bool
     {
         $cursor = $inlineContext->getCursor();
 
-        // The symbol must not have any other characters immediately prior
+        // The prefix must not have any other characters immediately prior
         $previousChar = $cursor->peek(-1);
         if ($previousChar !== null && $previousChar !== ' ') {
             // peek() doesn't modify the cursor, so no need to restore state first
@@ -69,8 +69,8 @@ final class MentionParser implements InlineParserInterface
         // Save the cursor state in case we need to rewind and bail
         $previousState = $cursor->saveState();
 
-        // Advance past the symbol to keep parsing simpler
-        $cursor->advanceBy(\mb_strlen($this->symbol));
+        // Advance past the prefix to keep parsing simpler
+        $cursor->advanceBy(\mb_strlen($this->prefix));
 
         // Parse the identifier
         $identifier = $cursor->match($this->mentionRegex);
@@ -81,7 +81,7 @@ final class MentionParser implements InlineParserInterface
             return false;
         }
 
-        $mention = $this->mentionGenerator->generateMention(new Mention($this->symbol, $identifier));
+        $mention = $this->mentionGenerator->generateMention(new Mention($this->prefix, $identifier));
 
         if ($mention === null) {
             $cursor->restoreState($previousState);
@@ -94,13 +94,13 @@ final class MentionParser implements InlineParserInterface
         return true;
     }
 
-    public static function createWithStringTemplate(string $symbol, string $mentionRegex, string $urlTemplate): MentionParser
+    public static function createWithStringTemplate(string $prefix, string $mentionRegex, string $urlTemplate): MentionParser
     {
-        return new self($symbol, $mentionRegex, new StringTemplateLinkGenerator($urlTemplate));
+        return new self($prefix, $mentionRegex, new StringTemplateLinkGenerator($urlTemplate));
     }
 
-    public static function createWithCallback(string $symbol, string $mentionRegex, callable $callback): MentionParser
+    public static function createWithCallback(string $prefix, string $mentionRegex, callable $callback): MentionParser
     {
-        return new self($symbol, $mentionRegex, new CallbackGenerator($callback));
+        return new self($prefix, $mentionRegex, new CallbackGenerator($callback));
     }
 }